+++ src/tribler-core/tribler_core/components/tests/test_tribler_components.py
@@ -47,46 +47,6 @@ def test_session_context_manager(loop, tribler_config):
         Session.current()
 
 
-async def test_ipv8_component(tribler_config):
-    tribler_config.ipv8.enabled = True
-    session = Session(tribler_config, [MasterKeyComponent(), RESTComponent(), Ipv8Component()])
-    with session:
-        await session.start()
-
-        comp = Ipv8Component.instance()
-        assert comp.started_event.is_set() and not comp.failed
-        assert comp.ipv8
-        assert comp.peer
-        assert not comp.dht_discovery_community
-        assert comp._task_manager
-        assert not comp._peer_discovery_community
-
-        await session.shutdown()
-
-
-async def test_ipv8_component_dht_disabled(tribler_config):
-    tribler_config.ipv8.enabled = True
-    tribler_config.dht.enabled = True
-    session = Session(tribler_config, [MasterKeyComponent(), RESTComponent(), Ipv8Component()])
-    with session:
-        await session.start()
-
-        comp = Ipv8Component.instance()
-        assert comp.dht_discovery_community
-
-
-async def test_ipv8_component_discovery_community_enabled(tribler_config):
-    tribler_config.ipv8.enabled = True
-    tribler_config.gui_test_mode = False
-    tribler_config.discovery_community.enabled = True
-    session = Session(tribler_config, [MasterKeyComponent(), RESTComponent(), Ipv8Component()])
-    with session:
-        await session.start()
-
-        comp = Ipv8Component.instance()
-        assert comp._peer_discovery_community
-
-
 async def test_libtorrent_component(tribler_config):
     tribler_config.libtorrent.enabled = True
     tribler_config.chant.enabled = True
